Refactoring Server and Core module
Solved cycle dependency problem in Core and Server modulle
