+++ Core/pom.xml
@@ -31,12 +31,5 @@
             <scope>compile</scope>
         </dependency>
 
-        <dependency>
-            <groupId>ValeraOnishkov.itmo.programing</groupId>
-            <artifactId>Server</artifactId>
-            <version>2.0</version>
-            <scope>compile</scope>
-        </dependency>
-
     </dependencies>
 </project>
