+++ Core/src/main/java/Commands/AverageOfEnginePower.java
@@ -2,9 +2,11 @@ package Commands;
 
 import Managers.Container;
 
+import java.nio.channels.SocketChannel;
+
 public class AverageOfEnginePower extends AbstractCommand {
     @Override
-    public Container execute(Container container) {
+    public Container execute(Container container, SocketChannel socketChannel) {
         return null;
 
     }
