+++ Core/src/main/java/Commands/Command.java
@@ -3,9 +3,11 @@ package Commands;
 import Managers.CollectionManager;
 import Managers.Container;
 
+import java.nio.channels.SocketChannel;
+
 public interface Command {
 
     CollectionManager collectionManager = new CollectionManager();
-    Container execute(Container container);
+    Container execute(Container container, SocketChannel socketChannel);
     String toString();
 }
