+++ Core/src/main/java/Managers/CollectionManager.java
@@ -5,20 +5,14 @@ import InputData.Vehicle;
 import java.io.IOException;
 import java.nio.channels.SocketChannel;
 import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Random;
+import java.util.Set;
+import java.util.stream.IntStream;
 
-public class SessionManager {
-
-    private static SocketChannel handlingChannel;
+public class CollectionManager {
 
     public static HashMap<SocketChannel, Session> sessions = new HashMap<>();
-
-    public static SocketChannel getHandlingChannel() {
-        return handlingChannel;
-    }
-
-    public static void setHandlingChannel(SocketChannel handlingChannel) {
-        SessionManager.handlingChannel = handlingChannel;
-    }
 
     public static HashMap<Integer, Vehicle> getSessionHashMap(SocketChannel socketChannel) {
         return sessions.get(socketChannel).getHashMap();
@@ -26,8 +20,8 @@ public class SessionManager {
     public static void setVariable(SocketChannel socketChannel, String path){
         sessions.get(socketChannel).initializeSession(path);
     }
-    public static void openSession(SocketChannel socketChannel, Session session){
-        sessions.put(socketChannel, session);
+    public static void openSession(SocketChannel socketChannel){
+        sessions.put(socketChannel, new Session());
     }
 
     public static void closeSession(SocketChannel socketChannel) throws IOException {
@@ -42,5 +36,26 @@ public class SessionManager {
     public static Session getSession(SocketChannel socketChannel){
         return sessions.get(socketChannel);
     }
+
+    public static Integer idGenerator(HashMap<Integer, Vehicle> hashMap){
+        Set<Integer> set = new HashSet<>();
+
+        for(Vehicle vehicle : hashMap.values()) {
+            set.add(vehicle.getId());
+        }
+
+        int[] numbers = IntStream.range(1, 2000000).filter(i -> !set.contains(i)).toArray();
+        int randomNumber = new Random().nextInt(numbers.length);
+
+        return numbers[randomNumber];
+    }
+
+
+
+
+
+
+
+
 
 }
