+++ Core/src/main/java/Managers/JsonHandler.java
