+++ Core/src/main/java/Managers/Session.java
