+++ Server/src/main/resources/Collection.json
@@ -1,0 +1,1 @@
+{}
